+++ Text/Earley/Parser.hs
@@ -2,9 +2,10 @@
 module Text.Earley.Parser
   ( Report(..)
   , Result(..)
+  , Parser
   , parser
   , allParses
   , fullParses
   , report
   ) where
-import Text.Earley.Internal
+import Text.Earley.Parser.Internal